fix : #7 senior 관련 클래스 폴더 이동 및 필드 추가
