--- src/main/java/org/farm/fireflyserver/care/domain/Gender.java
@@ -1,5 +1,0 @@
-package org.farm.fireflyserver.care.domain;
-
-public enum Gender {
-    MALE, FEMALE
-}
