+++ src/main/java/org/farm/fireflyserver/senior/domain/BenefitType.java
@@ -1,0 +1,15 @@
+package org.farm.fireflyserver.senior.domain;
+
+import lombok.Getter;
+import lombok.RequiredArgsConstructor;
+
+@Getter
+@RequiredArgsConstructor
+public enum BenefitType {
+    BASIC_LIVELIHOOD("기초수급자"),
+    BASIC_PENSION("기초연금"),
+    NEAR_POOR("차상위"),
+    OTHER("기타");
+
+    private final String description;
+}
