+++ src/main/java/org/farm/fireflyserver/senior/domain/Gender.java
@@ -1,0 +1,5 @@
+package org.farm.fireflyserver.senior.domain;
+
+public enum Gender {
+    MALE, FEMALE
+}
